requirements.txt
PyYAML==6.0.1
httpx>=0.27,<1
asyncpg>=0.29,<1
PyNaCl>=1.6,<2
